coding p01.

+++ eclipse-projects/projects/p01-motion-rc-vehicle/p01-motion-rc-vehicle-lib/src/org/programus/book/mobilelego/motion_rc_vehicle/comm/protocol/RobotReportMessage.java
@@ -3,25 +3,25 @@ package org.programus.book.mobilelego.motion_rc_vehicle.comm.protocol;
 public class RobotReportMessage implements NetMessage {
 	private static final long serialVersionUID = -8702695106516789834L;
 	
-	private int speed;
-	private int rotateSpeed;
-	private int distance;
-	public int getSpeed() {
+	private double speed;
+	private float rotateSpeed;
+	private double distance;
+	public double getSpeed() {
 		return speed;
 	}
-	public void setSpeed(int speed) {
+	public void setSpeed(double speed) {
 		this.speed = speed;
 	}
-	public int getRotateSpeed() {
+	public float getRotateSpeed() {
 		return rotateSpeed;
 	}
-	public void setRotateSpeed(int rotateSpeed) {
+	public void setRotateSpeed(float rotateSpeed) {
 		this.rotateSpeed = rotateSpeed;
 	}
-	public int getDistance() {
+	public double getDistance() {
 		return distance;
 	}
-	public void setDistance(int distance) {
+	public void setDistance(double distance) {
 		this.distance = distance;
 	}
 	@Override
